+++ src/transfers/DepositProvider.test.ts
@@ -94,35 +94,42 @@ describe("watchOneTransaction", () => {
   // suite-wide consts
   const transferServer = "https://www.stellar.org/transfers";
 
-  const pendingTransaction = (n: number = 0): { transaction: Transaction } => {
-    return {
-      transaction: {
-        kind: "deposit",
-        id: "TEST",
-        status: TransactionStatus.pending_anchor,
-        status_eta: n,
-      },
-    };
-  };
-  const successfulTransaction = (
-    n: number = 0,
+  const pendingTransaction = (
+    eta: number,
+    txStatus: TransactionStatus,
   ): { transaction: Transaction } => {
     return {
       transaction: {
         kind: "deposit",
         id: "TEST",
-        status: TransactionStatus.completed,
-        status_eta: n,
+        status: txStatus,
+        status_eta: eta,
       },
     };
   };
-  const failedTransaction = (n: number = 0): { transaction: Transaction } => {
+  const successfulTransaction = (
+    eta: number,
+    txStatus: TransactionStatus,
+  ): { transaction: Transaction } => {
     return {
       transaction: {
         kind: "deposit",
         id: "TEST",
-        status: TransactionStatus.error,
-        status_eta: n,
+        status: txStatus,
+        status_eta: eta,
+      },
+    };
+  };
+  const failedTransaction = (
+    eta: number,
+    txStatus: TransactionStatus,
+  ): { transaction: Transaction } => {
+    return {
+      transaction: {
+        kind: "deposit",
+        id: "TEST",
+        status: txStatus,
+        status_eta: eta,
       },
     };
   };
@@ -139,6 +146,7 @@ describe("watchOneTransaction", () => {
     provider = new DepositProvider(
       transferServer,
       StellarSdk.Keypair.random().publicKey(),
+      "uk-UA",
     );
     provider.authToken = "test";
 
@@ -149,7 +157,7 @@ describe("watchOneTransaction", () => {
     clock.restore();
   });
 
-  test("One success", async (done) => {
+  test("One completed success", async (done) => {
     const onMessage = sinon.spy((transaction) => {
       done(`onMessage incorrectly called with ${JSON.stringify(transaction)}`);
     });
@@ -162,16 +170,19 @@ describe("watchOneTransaction", () => {
 
     // queue up a success
     // @ts-ignore
-    fetch.mockResponses([JSON.stringify(successfulTransaction())]);
+    fetch.mockResponses([
+      JSON.stringify(successfulTransaction(0, TransactionStatus.completed)),
+    ]);
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: successfulTransaction(0, TransactionStatus.completed).transaction.id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -183,7 +194,44 @@ describe("watchOneTransaction", () => {
     clock.next();
   });
 
-  test("One pending message", async (done) => {
+  test("One refunded success", async (done) => {
+    const onMessage = sinon.spy((transaction) => {
+      done(`onMessage incorrectly called with ${JSON.stringify(transaction)}`);
+    });
+    const onSuccess = sinon.spy(() => {
+      done();
+    });
+    const onError = sinon.spy((e) => {
+      done(`onError incorrectly called with ${e.toString}`);
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify(successfulTransaction(0, TransactionStatus.refunded)),
+    ]);
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: successfulTransaction(0, TransactionStatus.refunded).transaction.id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    // wait a second, then done will call back onSuccess
+    clock.next();
+  });
+
+  test("One pending_user_transfer_start message", async (done) => {
     const onMessage = sinon.spy(() => {
       done();
     });
@@ -196,16 +244,24 @@ describe("watchOneTransaction", () => {
 
     // queue up a success
     // @ts-ignore
-    fetch.mockResponses([JSON.stringify(pendingTransaction(0))]);
+    fetch.mockResponses([
+      JSON.stringify(
+        pendingTransaction(0, TransactionStatus.pending_user_transfer_start),
+      ),
+    ]);
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: successfulTransaction(
+        0,
+        TransactionStatus.pending_user_transfer_start,
+      ).transaction.id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -213,7 +269,87 @@ describe("watchOneTransaction", () => {
     expect(onSuccess.callCount).toBe(0);
     expect(onError.callCount).toBe(0);
 
-    // wait a second, then done will call back success
+    // wait a second, then done will call back onMessage
+    clock.next();
+  });
+
+  test("One pending_user_transfer_complete message", async (done) => {
+    const onMessage = sinon.spy(() => {
+      done();
+    });
+    const onSuccess = sinon.spy((transaction) => {
+      done(`onSuccess incorrectly called with ${JSON.stringify(transaction)}`);
+    });
+    const onError = sinon.spy((e) => {
+      done(`onError incorrectly called with ${JSON.stringify(e)}`);
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify(
+        pendingTransaction(0, TransactionStatus.pending_user_transfer_complete),
+      ),
+    ]);
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: successfulTransaction(
+        0,
+        TransactionStatus.pending_user_transfer_complete,
+      ).transaction.id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    // wait a second, then done will call back onMessage
+    clock.next();
+  });
+
+  test("One pending_anchor message", async (done) => {
+    const onMessage = sinon.spy(() => {
+      done();
+    });
+    const onSuccess = sinon.spy((transaction) => {
+      done(`onSuccess incorrectly called with ${JSON.stringify(transaction)}`);
+    });
+    const onError = sinon.spy((e) => {
+      done(`onError incorrectly called with ${JSON.stringify(e)}`);
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify(pendingTransaction(0, TransactionStatus.pending_anchor)),
+    ]);
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: successfulTransaction(0, TransactionStatus.pending_anchor).transaction
+        .id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    // wait a second, then done will call back onMessage
     clock.next();
   });
 
@@ -230,16 +366,19 @@ describe("watchOneTransaction", () => {
 
     // then, queue up a success
     // @ts-ignore
-    fetch.mockResponses([JSON.stringify(failedTransaction(0))]);
+    fetch.mockResponses([
+      JSON.stringify(failedTransaction(0, TransactionStatus.error)),
+    ]);
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: successfulTransaction(0, TransactionStatus.error).transaction.id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -247,13 +386,50 @@ describe("watchOneTransaction", () => {
     expect(onSuccess.callCount).toBe(0);
     expect(onError.callCount).toBe(0);
 
-    // wait a second, then done will call back
+    // wait a second, then done will call back onError
+    clock.next();
+  });
+
+  test("One no_market", async (done) => {
+    const onMessage = sinon.spy((transaction) => {
+      done(`onMessage incorrectly called with ${JSON.stringify(transaction)}`);
+    });
+    const onSuccess = sinon.spy((transaction) => {
+      done(`onSuccess incorrectly called with ${JSON.stringify(transaction)}`);
+    });
+    const onError = sinon.spy(() => {
+      done();
+    });
+
+    // then, queue up a success
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify(failedTransaction(0, TransactionStatus.no_market)),
+    ]);
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: successfulTransaction(0, TransactionStatus.no_market).transaction.id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    // wait a second, then done will call back onError
     clock.next();
   });
 
   test("Several pending transactions", async (done) => {
     const onMessage = sinon.spy(() => {
-      expect(onMessage.callCount).toBeLessThan(5);
+      expect(onMessage.callCount).toBeLessThan(8);
     });
     const onSuccess = sinon.spy((transaction) => {
       done(`onSuccess incorrectly called with ${JSON.stringify(transaction)}`);
@@ -265,20 +441,57 @@ describe("watchOneTransaction", () => {
     // queue up a success
     // @ts-ignore
     fetch.mockResponses(
-      [JSON.stringify(pendingTransaction(0)), { status: 200 }],
-      [JSON.stringify(pendingTransaction(1)), { status: 200 }],
-      [JSON.stringify(pendingTransaction(2)), { status: 200 }],
-      [JSON.stringify(pendingTransaction(3)), { status: 200 }],
+      [
+        JSON.stringify(pendingTransaction(0, TransactionStatus.pending_anchor)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(
+          pendingTransaction(1, TransactionStatus.pending_external),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(
+          pendingTransaction(2, TransactionStatus.pending_stellar),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(pendingTransaction(3, TransactionStatus.pending_trust)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(pendingTransaction(4, TransactionStatus.pending_user)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(
+          pendingTransaction(
+            5,
+            TransactionStatus.pending_user_transfer_complete,
+          ),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(
+          pendingTransaction(6, TransactionStatus.pending_user_transfer_start),
+        ),
+        { status: 200 },
+      ],
     );
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: successfulTransaction(0, TransactionStatus.pending_anchor).transaction
+        .id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -286,7 +499,16 @@ describe("watchOneTransaction", () => {
     expect(onSuccess.callCount).toBe(0);
     expect(onError.callCount).toBe(0);
 
-    // wait a second, then done will call back success
+    // wait a second, then done will call back onMessage
+
+    clock.next();
+    await sleep(10);
+
+    clock.next();
+    await sleep(10);
+
+    clock.next();
+    await sleep(10);
 
     clock.next();
     await sleep(10);
@@ -303,7 +525,7 @@ describe("watchOneTransaction", () => {
     done();
   });
 
-  test("One pending, one success, no more after that", async () => {
+  test("One pending, one completed, no more after that", async () => {
     const onMessage = sinon.spy(() => {
       expect(onMessage.callCount).toBeLessThan(2);
     });
@@ -317,20 +539,116 @@ describe("watchOneTransaction", () => {
     // queue up a success
     // @ts-ignore
     fetch.mockResponses(
-      [JSON.stringify(pendingTransaction()), { status: 200 }],
-      [JSON.stringify(successfulTransaction()), { status: 200 }],
-      [JSON.stringify(successfulTransaction()), { status: 200 }],
-      [JSON.stringify(successfulTransaction()), { status: 200 }],
+      [
+        JSON.stringify(
+          pendingTransaction(
+            0,
+            TransactionStatus.pending_user_transfer_complete,
+          ),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.completed)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.completed)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.completed)),
+        { status: 200 },
+      ],
     );
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: successfulTransaction(0, TransactionStatus.completed).transaction.id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    clock.next();
+    await sleep(10);
+
+    // wait a second, then the pending should resolve
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    clock.next();
+    await sleep(10);
+
+    // the second time, a success should happen
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(1);
+    expect(onError.callCount).toBe(0);
+
+    clock.next();
+    await sleep(10);
+
+    // the third time, nothing should change or run again
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(1);
+    expect(onError.callCount).toBe(0);
+  });
+
+  test("One pending, one refunded, no more after that", async () => {
+    const onMessage = sinon.spy(() => {
+      expect(onMessage.callCount).toBeLessThan(2);
+    });
+    const onSuccess = sinon.spy(() => {
+      expect(onMessage.callCount).toBeLessThan(2);
+    });
+    const onError = sinon.spy((transaction) => {
+      expect(transaction).toBeUndefined();
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses(
+      [
+        JSON.stringify(
+          pendingTransaction(
+            0,
+            TransactionStatus.pending_user_transfer_complete,
+          ),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.refunded)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.refunded)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(successfulTransaction(0, TransactionStatus.refunded)),
+        { status: 200 },
+      ],
+    );
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: successfulTransaction(0, TransactionStatus.refunded).transaction.id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -377,20 +695,34 @@ describe("watchOneTransaction", () => {
     // queue up a success
     // @ts-ignore
     fetch.mockResponses(
-      [JSON.stringify(pendingTransaction()), { status: 200 }],
-      [JSON.stringify(failedTransaction()), { status: 200 }],
-      [JSON.stringify(failedTransaction()), { status: 200 }],
-      [JSON.stringify(failedTransaction()), { status: 200 }],
+      [
+        JSON.stringify(pendingTransaction(0, TransactionStatus.pending_anchor)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.error)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.error)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.error)),
+        { status: 200 },
+      ],
     );
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: pendingTransaction(0, TransactionStatus.pending_anchor).transaction
+        .id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -409,7 +741,81 @@ describe("watchOneTransaction", () => {
     clock.next();
     await sleep(10);
 
-    // the second time, a success should happen
+    // the second time, an error should happen
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(1);
+
+    clock.next();
+    await sleep(10);
+
+    // the third time, nothing should change or run again
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(1);
+  });
+
+  test("One pending, one no_market, no more after that", async () => {
+    const onMessage = sinon.spy(() => {
+      expect(onMessage.callCount).toBeLessThan(2);
+    });
+    const onSuccess = sinon.spy((transaction) => {
+      expect(transaction).toBeUndefined();
+    });
+    const onError = sinon.spy(() => {
+      expect(onError.callCount).toBeLessThan(2);
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses(
+      [
+        JSON.stringify(pendingTransaction(0, TransactionStatus.pending_anchor)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.no_market)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.no_market)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(0, TransactionStatus.no_market)),
+        { status: 200 },
+      ],
+    );
+
+    // start watching
+    provider.watchOneTransaction({
+      asset_code: "SMX",
+      id: pendingTransaction(0, TransactionStatus.pending_anchor).transaction
+        .id,
+      onMessage,
+      onSuccess,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    clock.next();
+    await sleep(10);
+
+    // wait a second, then the pending should resolve
+    expect(onMessage.callCount).toBe(1);
+    expect(onSuccess.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    clock.next();
+    await sleep(10);
+
+    // the second time, an error should happen
     expect(onMessage.callCount).toBe(1);
     expect(onSuccess.callCount).toBe(0);
     expect(onError.callCount).toBe(1);
@@ -437,21 +843,45 @@ describe("watchOneTransaction", () => {
     // queue up a success
     // @ts-ignore
     fetch.mockResponses(
-      [JSON.stringify(pendingTransaction(0)), { status: 200 }],
-      [JSON.stringify(pendingTransaction(1)), { status: 200 }],
-      [JSON.stringify(failedTransaction(2)), { status: 200 }],
-      [JSON.stringify(failedTransaction(3)), { status: 200 }],
-      [JSON.stringify(failedTransaction(4)), { status: 200 }],
+      [
+        JSON.stringify(
+          pendingTransaction(0, TransactionStatus.pending_user_transfer_start),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(
+          pendingTransaction(
+            1,
+            TransactionStatus.pending_user_transfer_complete,
+          ),
+        ),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(2, TransactionStatus.error)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(3, TransactionStatus.error)),
+        { status: 200 },
+      ],
+      [
+        JSON.stringify(failedTransaction(4, TransactionStatus.error)),
+        { status: 200 },
+      ],
     );
 
     // start watching
     provider.watchOneTransaction({
       asset_code: "SMX",
-      id: successfulTransaction(0).transaction.id,
+      id: pendingTransaction(0, TransactionStatus.pending_user_transfer_start)
+        .transaction.id,
       onMessage,
       onSuccess,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -512,6 +942,7 @@ describe("watchAllTransactions", () => {
     provider = new DepositProvider(
       transferServer,
       StellarSdk.Keypair.random().publicKey(),
+      "uk-UA",
     );
 
     provider.authToken = "test";
@@ -524,37 +955,6 @@ describe("watchAllTransactions", () => {
   });
 
   test("Return only pending", async () => {
-    const onMessage = sinon.spy(() => {
-      expect(onMessage.callCount).toBeLessThan(3);
-    });
-
-    const onError = sinon.spy((e) => {
-      expect(e).toBeUndefined();
-    });
-
-    // queue up a success
-    // @ts-ignore
-    fetch.mockResponses([JSON.stringify(TransactionsResponse)]);
-
-    // start watching
-    provider.watchAllTransactions({
-      asset_code: "SMX",
-      onMessage,
-      onError,
-      timeout: 10,
-    });
-
-    // nothing should run at first
-    expect(onMessage.callCount).toBe(0);
-    expect(onError.callCount).toBe(0);
-
-    await sleep(1);
-
-    expect(onMessage.callCount).toBe(2);
-    expect(onError.callCount).toBe(0);
-  });
-
-  test("Return one changed thing", async () => {
     const onMessage = sinon.spy(() => {
       expect(onMessage.callCount).toBeLessThan(4);
     });
@@ -573,6 +973,7 @@ describe("watchAllTransactions", () => {
       onMessage,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -581,33 +982,97 @@ describe("watchAllTransactions", () => {
 
     await sleep(1);
 
-    expect(onMessage.callCount).toBe(2);
+    expect(onMessage.callCount).toBe(3);
+    expect(onError.callCount).toBe(0);
+  });
+
+  test("Return two changed thing", async () => {
+    const onMessage = sinon.spy(() => {
+      expect(onMessage.callCount).toBeLessThan(6);
+    });
+
+    const onError = sinon.spy((e) => {
+      expect(e).toBeUndefined();
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses([JSON.stringify(TransactionsResponse)]);
+
+    // start watching
+    provider.watchAllTransactions({
+      asset_code: "SMX",
+      onMessage,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
     expect(onError.callCount).toBe(0);
 
-    // change one thing to "Successful"
-    const [firstResponse, ...rest] = TransactionsResponse.transactions;
+    await sleep(1);
+
+    expect(onMessage.callCount).toBe(3);
+    expect(onError.callCount).toBe(0);
+
+    // change one thing to "completed"
+    const [firstResponseA, ...restA] = TransactionsResponse.transactions;
+    let updatedTransactions = [
+      {
+        ...firstResponseA,
+        status: TransactionStatus.completed,
+      },
+      ...restA,
+    ];
+
     // @ts-ignore
     fetch.mockResponses([
       JSON.stringify({
         ...TransactionsResponse,
-        transactions: [
-          {
-            ...firstResponse,
-            status: TransactionStatus.completed,
-          },
-          ...rest,
-        ],
+        transactions: updatedTransactions,
       }),
     ]);
 
     clock.next();
     await sleep(10);
 
-    expect(onMessage.callCount).toBe(3);
+    expect(onMessage.callCount).toBe(4);
+    expect(onError.callCount).toBe(0);
+
+    await sleep(1);
+
+    expect(onMessage.callCount).toBe(4);
+    expect(onError.callCount).toBe(0);
+
+    // change another thing to "refunded"
+    const [firstResponseB, secondResponseB, ...restB] = updatedTransactions;
+    updatedTransactions = [
+      firstResponseB,
+      {
+        ...secondResponseB,
+        status: TransactionStatus.refunded,
+      },
+      ...restB,
+    ];
+
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify({
+        ...TransactionsResponse,
+        transactions: updatedTransactions,
+      }),
+    ]);
+
+    clock.next();
+    await sleep(10);
+
+    expect(onMessage.callCount).toBe(5);
     expect(onError.callCount).toBe(0);
   });
 
-  test("Immediate successes get messages", async () => {
+  test("Immediate completed get messages", async () => {
     const onMessage = sinon.spy(() => {
       expect(onMessage.callCount).toBeLessThan(2);
     });
@@ -631,6 +1096,7 @@ describe("watchAllTransactions", () => {
       onMessage,
       onError,
       timeout: 10,
+      lang: "uk-UA",
     });
 
     // nothing should run at first
@@ -716,6 +1182,117 @@ describe("watchAllTransactions", () => {
     expect(onMessage.callCount).toBe(1);
     expect(onError.callCount).toBe(0);
   });
+
+  test("Immediate refunded get messages", async () => {
+    const onMessage = sinon.spy(() => {
+      expect(onMessage.callCount).toBeLessThan(2);
+    });
+
+    const onError = sinon.spy((e) => {
+      expect(e).toBeUndefined();
+    });
+
+    // queue up a success
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify({
+        status: true,
+        transactions: [],
+      }),
+    ]);
+
+    // start watching
+    provider.watchAllTransactions({
+      asset_code: "SMX",
+      onMessage,
+      onError,
+      timeout: 10,
+      lang: "uk-UA",
+    });
+
+    // nothing should run at first
+    expect(onMessage.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    await sleep(1);
+
+    // still nothing
+    expect(onMessage.callCount).toBe(0);
+    expect(onError.callCount).toBe(0);
+
+    // add a new success message
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify({
+        status: true,
+        transactions: [
+          {
+            id: 77,
+            order_id: "ord_2mRPdvpUjPN2rfuS7",
+            status: "refunded",
+            receiving_account_number: "646180111812345678",
+            receiving_account_bank: "STP",
+            memo: null,
+            memo_type: null,
+            email_address: null,
+            type: "SPEI",
+            asset_code: "SMX",
+            expires_at: "1970-01-19T09:23:30.864Z",
+            created_at: "2019-10-01T20:34:24.753Z",
+            phone: "+14155099007",
+            amount: "25.00",
+            account: "GARBRF35ZWKO4MIEY7RPHHPQ74P45TWNMIQ5VKOTXF7KYVFJBGWW2IMQ",
+            transaction_id: "DI938432727A0B2456",
+            external_transaction_id: "DI938432727A0B2456",
+            kind: "deposit",
+          },
+        ],
+      }),
+    ]);
+
+    clock.next();
+    await sleep(10);
+
+    // should have a success
+    expect(onMessage.callCount).toBe(1);
+    expect(onError.callCount).toBe(0);
+
+    // getting the same thing again should change nothing
+    // @ts-ignore
+    fetch.mockResponses([
+      JSON.stringify({
+        status: true,
+        transactions: [
+          {
+            id: 77,
+            order_id: "ord_2mRPdvpUjPN2rfuS7",
+            status: "refunded",
+            receiving_account_number: "646180111812345678",
+            receiving_account_bank: "STP",
+            memo: null,
+            memo_type: null,
+            email_address: null,
+            type: "SPEI",
+            asset_code: "SMX",
+            expires_at: "1970-01-19T09:23:30.864Z",
+            created_at: "2019-10-01T20:34:24.753Z",
+            phone: "+14155099007",
+            amount: "25.00",
+            account: "GARBRF35ZWKO4MIEY7RPHHPQ74P45TWNMIQ5VKOTXF7KYVFJBGWW2IMQ",
+            transaction_id: "DI938432727A0B2456",
+            external_transaction_id: "DI938432727A0B2456",
+            kind: "deposit",
+          },
+        ],
+      }),
+    ]);
+
+    clock.next();
+    await sleep(10);
+
+    expect(onMessage.callCount).toBe(1);
+    expect(onError.callCount).toBe(0);
+  });
 });
 
 describe("validateFields", () => {
@@ -725,6 +1302,7 @@ describe("validateFields", () => {
     const provider = new DepositProvider(
       "https://test.com",
       StellarSdk.Keypair.random().publicKey(),
+      "en-US",
     );
     provider.info = {
       deposit: {